Converted package bdema version BLP_LIB_BDE_BDE_2.5.0 from clearcase

+++ groups/bde/bdema/bdema_buffermanager.h
@@ -461,7 +461,7 @@ void *bdema_BufferManager::allocate(bsls_PlatformUtil::size_type size)
     return bdema_BufferImpUtil::allocateFromBuffer(&d_cursor,
                                                    d_buffer_p,
                                                    d_bufferSize,
-                                                   size,
+                                                   static_cast<int>(size),
                                                    d_alignmentStrategy);
 }
